US 10.2 Coverage Plugin: Changed coverage counter from line to instructions; improved naming in unit tests.
